requirements.txt
torch
torchvision
opencv-python
pandas
numpy
scikit-learn
scikit-image
matplotlib
openpyxl
Pillow